Write form and table components
Define new React components that accept user input and display the
amortization information calculated from the user-supplied values.

The app is very basic at this point, only accepting four pieces of
information from the user and displaying the monthly mortgage amounts
for the entire amortization period. In the future, the app should
allow the user to specify the term of the mortgage and only display
the amounts for the span of that term (i.e., a 5-year term). The app
should also allow the user to calculate the return on their
investment.